Introduce TableScan.with_shard
