--- paimon-python/pypaimon/tests/rest_catalog_test.py
@@ -1,479 +1,0 @@
-"""
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-"""
-import glob
-import logging
-import os
-import shutil
-import tempfile
-import unittest
-import uuid
-
-import pandas as pd
-import pyarrow as pa
-
-from pypaimon.api import ConfigResponse, Identifier
-from pypaimon.api.auth import BearTokenAuthProvider
-from pypaimon.api.options import Options
-from pypaimon.catalog.catalog_context import CatalogContext
-from pypaimon.catalog.catalog_factory import CatalogFactory
-from pypaimon.catalog.rest.rest_catalog import RESTCatalog
-from pypaimon.catalog.table_metadata import TableMetadata
-from pypaimon.schema.data_types import (ArrayType, AtomicType, DataField,
-                                        MapType)
-from pypaimon.schema.schema import Schema
-from pypaimon.schema.table_schema import TableSchema
-from pypaimon.tests.rest_server import RESTCatalogServer
-
-
-class RESTCatalogTest(unittest.TestCase):
-    def setUp(self):
-        self.temp_dir = tempfile.mkdtemp(prefix="unittest_")
-        self.warehouse = os.path.join(self.temp_dir, 'warehouse')
-
-        self.config = ConfigResponse(defaults={"prefix": "mock-test"})
-        self.token = str(uuid.uuid4())
-        self.server = RESTCatalogServer(
-            data_path=self.temp_dir,
-            auth_provider=BearTokenAuthProvider(self.token),
-            config=self.config,
-            warehouse="warehouse"
-        )
-        self.server.start()
-        print(f"\nServer started at: {self.server.get_url()}")
-
-        self.options = {
-            'metastore': 'rest',
-            'uri': f"http://localhost:{self.server.port}",
-            'warehouse': "warehouse",
-            'dlf.region': 'cn-hangzhou',
-            "token.provider": "bear",
-            'token': self.token,
-        }
-        self.rest_catalog = CatalogFactory.create(self.options)
-        self.rest_catalog.create_database("default", False)
-
-        self.pa_schema = pa.schema([
-            ('user_id', pa.int64()),
-            ('item_id', pa.int64()),
-            ('behavior', pa.string()),
-            ('dt', pa.string()),
-            ('long-dt', pa.string())
-        ])
-        self.raw_data = {
-            'user_id': [1, 2, 3, 4, 5, 6, 7, 8],
-            'item_id': [1001, 1002, 1003, 1004, 1005, 1006, 1007, 1008],
-            'behavior': ['a', 'b', 'c', None, 'e', 'f', 'g', 'h'],
-            'dt': ['p1', 'p1', 'p2', 'p1', 'p2', 'p1', 'p2', 'p2'],
-            'long-dt': ['2024-10-10', '2024-10-10', '2024-10-10', '2024-01-01', '2024-10-10', '2025-01-23',
-                        'abcdefghijklmnopk', '2025-08-08']
-        }
-        self.expected = pa.Table.from_pydict(self.raw_data, schema=self.pa_schema)
-
-        schema = Schema.from_pyarrow_schema(self.pa_schema)
-        self.rest_catalog.create_table('default.test_reader_iterator', schema, False)
-        self.table = self.rest_catalog.get_table('default.test_reader_iterator')
-        write_builder = self.table.new_batch_write_builder()
-        table_write = write_builder.new_write()
-        table_commit = write_builder.new_commit()
-        table_write.write_arrow(self.expected)
-        table_commit.commit(table_write.prepare_commit())
-        table_write.close()
-        table_commit.close()
-
-    def tearDown(self):
-        # Shutdown server
-        self.server.shutdown()
-        print("Server stopped")
-        shutil.rmtree(self.temp_dir, ignore_errors=True)
-
-    def test_rest_catalog(self):
-        """Example usage of RESTCatalogServer"""
-        # Setup logging
-        logging.basicConfig(level=logging.INFO)
-
-        # Create config
-        config = ConfigResponse(defaults={"prefix": "mock-test"})
-        token = str(uuid.uuid4())
-        # Create server
-        server = RESTCatalogServer(
-            data_path="/tmp/test_warehouse",
-            auth_provider=BearTokenAuthProvider(token),
-            config=config,
-            warehouse="test_warehouse"
-        )
-        try:
-            # Start server
-            server.start()
-            print(f"Server started at: {server.get_url()}")
-            test_databases = {
-                "default": server.mock_database("default", {"env": "test"}),
-                "test_db1": server.mock_database("test_db1", {"env": "test"}),
-                "test_db2": server.mock_database("test_db2", {"env": "test"}),
-                "prod_db": server.mock_database("prod_db", {"env": "prod"})
-            }
-            data_fields = [
-                DataField(0, "name", AtomicType('INT'), 'desc  name'),
-                DataField(1, "arr11", ArrayType(True, AtomicType('INT')), 'desc  arr11'),
-                DataField(2, "map11", MapType(False, AtomicType('INT'),
-                                              MapType(False, AtomicType('INT'), AtomicType('INT'))),
-                          'desc  arr11'),
-            ]
-            schema = TableSchema(TableSchema.CURRENT_VERSION, len(data_fields), data_fields, len(data_fields),
-                                 [], [], {}, "")
-            test_tables = {
-                "default.user": TableMetadata(uuid=str(uuid.uuid4()), is_external=True, schema=schema),
-            }
-            server.table_metadata_store.update(test_tables)
-            server.database_store.update(test_databases)
-            options = {
-                'uri': f"http://localhost:{server.port}",
-                'warehouse': 'test_warehouse',
-                'dlf.region': 'cn-hangzhou',
-                "token.provider": "bear",
-                'token': token
-            }
-            rest_catalog = RESTCatalog(CatalogContext.create_from_options(Options(options)))
-            self.assertSetEqual(set(rest_catalog.list_databases()), {*test_databases})
-            self.assertEqual(rest_catalog.get_database('default').name, test_databases.get('default').name)
-            table = rest_catalog.get_table(Identifier.from_string('default.user'))
-            self.assertEqual(table.identifier.get_full_name(), 'default.user')
-        finally:
-            # Shutdown server
-            server.shutdown()
-            print("Server stopped")
-
-    def test_write(self):
-        # Setup logging
-        logging.basicConfig(level=logging.INFO)
-
-        schema = Schema.from_pyarrow_schema(self.pa_schema, partition_keys=['dt'])
-        self.rest_catalog.create_table("default.test_table", schema, False)
-        table = self.rest_catalog.get_table("default.test_table")
-
-        data = {
-            'user_id': [1, 2, 3, 4],
-            'item_id': [1001, 1002, 1003, 1004],
-            'behavior': ['a', 'b', 'c', None],
-            'dt': ['p1', 'p1', 'p2', 'p1'],
-            'long-dt': ['2024-10-10', '2024-10-10', '2024-10-10', '2024-01-01']
-        }
-        expect = pa.Table.from_pydict(data, schema=self.pa_schema)
-
-        write_builder = table.new_batch_write_builder()
-        table_write = write_builder.new_write()
-        table_commit = write_builder.new_commit()
-        table_write.write_arrow(expect)
-        commit_messages = table_write.prepare_commit()
-        table_commit.commit(commit_messages)
-        table_write.close()
-        table_commit.close()
-
-        self.assertTrue(os.path.exists(self.warehouse + "/default/test_table/snapshot/LATEST"))
-        self.assertTrue(os.path.exists(self.warehouse + "/default/test_table/snapshot/snapshot-1"))
-        self.assertTrue(os.path.exists(self.warehouse + "/default/test_table/manifest"))
-        self.assertTrue(os.path.exists(self.warehouse + "/default/test_table/dt=p1"))
-        self.assertEqual(len(glob.glob(self.warehouse + "/default/test_table/manifest/*.avro")), 2)
-
-    def testParquetAppendOnlyReader(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema, partition_keys=['dt'])
-        self.rest_catalog.create_table('default.test_append_only_parquet', schema, False)
-        table = self.rest_catalog.get_table('default.test_append_only_parquet')
-        self._write_test_table(table)
-
-        read_builder = table.new_read_builder()
-        actual = self._read_test_table(read_builder).sort_by('user_id')
-        self.assertEqual(actual, self.expected)
-
-    def testOrcAppendOnlyReader(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema, partition_keys=['dt'], options={'file.format': 'orc'})
-        self.rest_catalog.create_table('default.test_append_only_orc', schema, False)
-        table = self.rest_catalog.get_table('default.test_append_only_orc')
-        self._write_test_table(table)
-
-        read_builder = table.new_read_builder()
-        actual = self._read_test_table(read_builder).sort_by('user_id')
-        self.assertEqual(actual, self.expected)
-
-    def testAvroAppendOnlyReader(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema, partition_keys=['dt'], options={'file.format': 'avro'})
-        self.rest_catalog.create_table('default.test_append_only_avro', schema, False)
-        table = self.rest_catalog.get_table('default.test_append_only_avro')
-        self._write_test_table(table)
-
-        read_builder = table.new_read_builder()
-        actual = self._read_test_table(read_builder).sort_by('user_id')
-        self.assertEqual(actual, self.expected)
-
-    def testAppendOnlyReaderWithFilter(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema, partition_keys=['dt'])
-        self.rest_catalog.create_table('default.test_append_only_filter', schema, False)
-        table = self.rest_catalog.get_table('default.test_append_only_filter')
-        self._write_test_table(table)
-
-        predicate_builder = table.new_read_builder().new_predicate_builder()
-        p1 = predicate_builder.less_than('user_id', 7)
-        p2 = predicate_builder.greater_or_equal('user_id', 2)
-        p3 = predicate_builder.between('user_id', 0, 6)  # [2/b, 3/c, 4/d, 5/e, 6/f] left
-        p4 = predicate_builder.is_not_in('behavior', ['b', 'e'])  # [3/c, 4/d, 6/f] left
-        p5 = predicate_builder.is_in('dt', ['p1'])  # exclude 3/c
-        p6 = predicate_builder.is_not_null('behavior')  # exclude 4/d
-        g1 = predicate_builder.and_predicates([p1, p2, p3, p4, p5, p6])
-        read_builder = table.new_read_builder().with_filter(g1)
-        actual = self._read_test_table(read_builder)
-        expected = pa.concat_tables([
-            self.expected.slice(5, 1)  # 6/f
-        ])
-        self.assertEqual(actual.sort_by('user_id'), expected)
-
-        p7 = predicate_builder.startswith('behavior', 'a')
-        p10 = predicate_builder.equal('item_id', 1002)
-        p11 = predicate_builder.is_null('behavior')
-        p9 = predicate_builder.contains('behavior', 'f')
-        p8 = predicate_builder.endswith('dt', 'p2')
-        g2 = predicate_builder.or_predicates([p7, p8, p9, p10, p11])
-        read_builder = table.new_read_builder().with_filter(g2)
-        actual = self._read_test_table(read_builder)
-        self.assertEqual(actual.sort_by('user_id'), self.expected)
-
-        g3 = predicate_builder.and_predicates([g1, g2])
-        read_builder = table.new_read_builder().with_filter(g3)
-        actual = self._read_test_table(read_builder)
-        expected = pa.concat_tables([
-            self.expected.slice(5, 1)  # 6/f
-        ])
-        self.assertEqual(actual.sort_by('user_id'), expected)
-
-        # Same as java, 'not_equal' will also filter records of 'None' value
-        p12 = predicate_builder.not_equal('behavior', 'f')
-        read_builder = table.new_read_builder().with_filter(p12)
-        actual = self._read_test_table(read_builder)
-        expected = pa.concat_tables([
-            # not only 6/f, but also 4/d will be filtered
-            self.expected.slice(0, 1),  # 1/a
-            self.expected.slice(1, 1),  # 2/b
-            self.expected.slice(2, 1),  # 3/c
-            self.expected.slice(4, 1),  # 5/e
-            self.expected.slice(6, 1),  # 7/g
-            self.expected.slice(7, 1),  # 8/h
-        ])
-        self.assertEqual(actual.sort_by('user_id'), expected)
-
-    def testAppendOnlyReaderWithProjection(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema, partition_keys=['dt'])
-        self.rest_catalog.create_table('default.test_append_only_projection', schema, False)
-        table = self.rest_catalog.get_table('default.test_append_only_projection')
-        self._write_test_table(table)
-
-        read_builder = table.new_read_builder().with_projection(['dt', 'user_id'])
-        actual = self._read_test_table(read_builder).sort_by('user_id')
-        expected = self.expected.select(['dt', 'user_id'])
-        self.assertEqual(actual, expected)
-
-    def testAvroAppendOnlyReaderWithProjection(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema, partition_keys=['dt'], options={'file.format': 'avro'})
-        self.rest_catalog.create_table('default.test_avro_append_only_projection', schema, False)
-        table = self.rest_catalog.get_table('default.test_avro_append_only_projection')
-        self._write_test_table(table)
-
-        read_builder = table.new_read_builder().with_projection(['dt', 'user_id'])
-        actual = self._read_test_table(read_builder).sort_by('user_id')
-        expected = self.expected.select(['dt', 'user_id'])
-        self.assertEqual(actual, expected)
-
-    def testAppendOnlyReaderWithLimit(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema, partition_keys=['dt'])
-        self.rest_catalog.create_table('default.test_append_only_limit', schema, False)
-        table = self.rest_catalog.get_table('default.test_append_only_limit')
-        self._write_test_table(table)
-
-        read_builder = table.new_read_builder().with_limit(1)
-        actual = self._read_test_table(read_builder)
-        # only records from 1st commit (1st split) will be read
-        # might be split of "dt=1" or split of "dt=2"
-        self.assertEqual(actual.num_rows, 4)
-
-    def testPkParquetReader(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema,
-                                            partition_keys=['dt'],
-                                            primary_keys=['user_id', 'dt'],
-                                            options={'bucket': '2'})
-        self.rest_catalog.create_table('default.test_pk_parquet', schema, False)
-        table = self.rest_catalog.get_table('default.test_pk_parquet')
-        self._write_test_table(table)
-
-        read_builder = table.new_read_builder()
-        actual = self._read_test_table(read_builder).sort_by('user_id')
-        self.assertEqual(actual, self.expected)
-
-    def testPkOrcReader(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema,
-                                            partition_keys=['dt'],
-                                            primary_keys=['user_id', 'dt'],
-                                            options={
-                                                'bucket': '1',
-                                                'file.format': 'orc'
-                                            })
-        self.rest_catalog.create_table('default.test_pk_orc', schema, False)
-        table = self.rest_catalog.get_table('default.test_pk_orc')
-        self._write_test_table(table)
-
-        read_builder = table.new_read_builder()
-        actual: pa.Table = self._read_test_table(read_builder).sort_by('user_id')
-
-        # when bucket=1, actual field name will contain 'not null', so skip comparing field name
-        for i in range(len(actual.columns)):
-            col_a = actual.column(i)
-            col_b = self.expected.column(i)
-            self.assertEqual(col_a, col_b)
-
-    def testPkAvroReader(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema,
-                                            partition_keys=['dt'],
-                                            primary_keys=['user_id', 'dt'],
-                                            options={
-                                                'bucket': '2',
-                                                'file.format': 'avro'
-                                            })
-        self.rest_catalog.create_table('default.test_pk_avro', schema, False)
-        table = self.rest_catalog.get_table('default.test_pk_avro')
-        self._write_test_table(table)
-
-        read_builder = table.new_read_builder()
-        actual = self._read_test_table(read_builder).sort_by('user_id')
-        self.assertEqual(actual, self.expected)
-
-    def testPkReaderWithFilter(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema,
-                                            partition_keys=['dt'],
-                                            primary_keys=['user_id', 'dt'],
-                                            options={'bucket': '2'})
-        self.rest_catalog.create_table('default.test_pk_filter', schema, False)
-        table = self.rest_catalog.get_table('default.test_pk_filter')
-        self._write_test_table(table)
-
-        predicate_builder = table.new_read_builder().new_predicate_builder()
-        p1 = predicate_builder.is_in('dt', ['p1'])
-        p2 = predicate_builder.between('user_id', 2, 7)
-        p3 = predicate_builder.is_not_null('behavior')
-        g1 = predicate_builder.and_predicates([p1, p2, p3])
-        read_builder = table.new_read_builder().with_filter(g1)
-        actual = self._read_test_table(read_builder).sort_by('user_id')
-        expected = pa.concat_tables([
-            self.expected.slice(1, 1),  # 2/b
-            self.expected.slice(5, 1)  # 7/g
-        ])
-        self.assertEqual(actual, expected)
-
-    def testPkReaderWithProjection(self):
-        schema = Schema.from_pyarrow_schema(self.pa_schema,
-                                            partition_keys=['dt'],
-                                            primary_keys=['user_id', 'dt'],
-                                            options={'bucket': '2'})
-        self.rest_catalog.create_table('default.test_pk_projection', schema, False)
-        table = self.rest_catalog.get_table('default.test_pk_projection')
-        self._write_test_table(table)
-
-        read_builder = table.new_read_builder().with_projection(['dt', 'user_id', 'behavior'])
-        actual = self._read_test_table(read_builder).sort_by('user_id')
-        expected = self.expected.select(['dt', 'user_id', 'behavior'])
-        self.assertEqual(actual, expected)
-
-    def testWriteWrongSchema(self):
-        self.rest_catalog.create_table('default.test_wrong_schema',
-                                       Schema.from_pyarrow_schema(self.pa_schema),
-                                       False)
-        table = self.rest_catalog.get_table('default.test_wrong_schema')
-
-        data = {
-            'f0': [1, 2, 3],
-            'f1': ['a', 'b', 'c'],
-        }
-        df = pd.DataFrame(data)
-        schema = pa.schema([
-            ('f0', pa.int64()),
-            ('f1', pa.string())
-        ])
-        record_batch = pa.RecordBatch.from_pandas(df, schema)
-
-        write_builder = table.new_batch_write_builder()
-        table_write = write_builder.new_write()
-
-        with self.assertRaises(ValueError) as e:
-            table_write.write_arrow_batch(record_batch)
-        self.assertTrue(str(e.exception).startswith("Input schema isn't consistent with table schema."))
-
-    def testReaderIterator(self):
-        read_builder = self.table.new_read_builder()
-        table_read = read_builder.new_read()
-        splits = read_builder.new_scan().plan().splits()
-        iterator = table_read.to_iterator(splits)
-        result = []
-        value = next(iterator, None)
-        while value is not None:
-            result.append(value.get_field(1))
-            value = next(iterator, None)
-        self.assertEqual(result, [1001, 1002, 1003, 1004, 1005, 1006, 1007, 1008])
-
-    def testReaderDuckDB(self):
-        read_builder = self.table.new_read_builder()
-        table_read = read_builder.new_read()
-        splits = read_builder.new_scan().plan().splits()
-        duckdb_con = table_read.to_duckdb(splits, 'duckdb_table')
-        actual = duckdb_con.query("SELECT * FROM duckdb_table").fetchdf()
-        expect = pd.DataFrame(self.raw_data)
-        pd.testing.assert_frame_equal(actual.reset_index(drop=True), expect.reset_index(drop=True))
-
-    def _write_test_table(self, table):
-        write_builder = table.new_batch_write_builder()
-
-        # first write
-        table_write = write_builder.new_write()
-        table_commit = write_builder.new_commit()
-        data1 = {
-            'user_id': [1, 2, 3, 4],
-            'item_id': [1001, 1002, 1003, 1004],
-            'behavior': ['a', 'b', 'c', None],
-            'dt': ['p1', 'p1', 'p2', 'p1'],
-            'long-dt': ['2024-10-10', '2024-10-10', '2024-10-10', '2024-01-01'],
-        }
-        pa_table = pa.Table.from_pydict(data1, schema=self.pa_schema)
-        table_write.write_arrow(pa_table)
-        table_commit.commit(table_write.prepare_commit())
-        table_write.close()
-        table_commit.close()
-
-        # second write
-        table_write = write_builder.new_write()
-        table_commit = write_builder.new_commit()
-        data2 = {
-            'user_id': [5, 6, 7, 8],
-            'item_id': [1005, 1006, 1007, 1008],
-            'behavior': ['e', 'f', 'g', 'h'],
-            'dt': ['p2', 'p1', 'p2', 'p2'],
-            'long-dt': ['2024-10-10', '2025-01-23', 'abcdefghijklmnopk', '2025-08-08'],
-        }
-        pa_table = pa.Table.from_pydict(data2, schema=self.pa_schema)
-        table_write.write_arrow(pa_table)
-        table_commit.commit(table_write.prepare_commit())
-        table_write.close()
-        table_commit.close()
-
-    def _read_test_table(self, read_builder):
-        table_read = read_builder.new_read()
-        splits = read_builder.new_scan().plan().splits()
-        return table_read.to_arrow(splits)
